Run csvs-to-sqlite in Travis, then publish to Now
I want to deploy a fresh instance to Now on the following:

* The FiveThirtyEight data has changed in some way
* Datasette master has changed in some way

Using 'RUN wget ...' in the Dockerfile does not achieve this, as those lines
will be layer cached whether or not the wgetted content has changed.

Instead, I'm having Travis CI download the files and build them into the
fivethirtyeight.db and metadata.json

I can then trust Now to notice if the contents of those files have changed,
and trigger a new build.

I also download Datasette master as a zip archive and send that to Now. This
means that if the archive contents has changed Now should deploy a new build.

Small functionality change: source_url for tables should now point to the
original CSV in the FiveThirtyEight GitHub repo.